Move `fillConstantsInString` function from `src` folder to `src/operation/utils`
